more controller refactoring
