+++ backend/src/Controllers/getElectionResultsController.ts
@@ -1,0 +1,42 @@
+import ServiceLocator from "../ServiceLocator";
+import Logger from "../Services/Logging/Logger";
+import { BadRequest } from "@curveball/http-errors";
+import { Ballot } from '../../../domain_model/Ballot';
+import { Score } from '../../../domain_model/Score';
+
+const StarResults = require('../Tabulators/StarResults.js');
+const BallotModel = ServiceLocator.ballotsDb();
+
+const className = "VoterRolls.Controllers";
+
+const getElectionResults = async (req: any, res: any, next: any) => {
+    var electionId = req.election.election_id;
+    Logger.info(req, `${className}.getElectionResults: ${electionId}`);
+    const ballots = await BallotModel.getBallotsByElectionID(String(electionId), req);
+    if (!ballots) {
+        const msg = `Ballots not found for Election ${electionId}`;
+        Logger.info(req, msg);
+        throw new BadRequest(msg);
+    }
+
+    const election = req.election
+    const candidateNames = election.races[0].candidates.map((Candidate: any) => (Candidate.candidate_name))
+    const cvr = ballots.map((ballot: Ballot) => (
+        ballot.votes[0].scores.map((score: Score) => (
+            score.score
+        ))
+    ))
+    const num_winners = election.races[0].num_winners
+    const results = StarResults(candidateNames, cvr, num_winners)
+
+    res.json(
+        {
+            Election: election,
+            Results: results
+        }
+    )
+}
+
+module.exports = {
+    getElectionResults
+}
